Add files via upload
This is more or less the same file as before, I have just removed emojis and rows that contain one or more occurrences of '@user' and nothing else. There is a short tutorial on the use of literal_eval, please use it if necessary.
Added on 17th April 2020 at 8:30 a.m.