Add files via upload
Label 2 Analysis using Naïve Bayes Classifier, Logistic Regression, SVM, KNN and Multi-layer Perceptron. It also includes an analysis of KNN for different nearest neighbour values (the optimal being 9 neighbours). It also incorporates an analysis of the optimal value of the Regularization Parameter and the Number of Hidden Layers by varying them over a range of respective values in the case of MLP Classifier. Optimal values: alpha = 0.001 and Number of Hidden Layers = 10